Move the config discovery into common module
Since there already exists a common module that exposes functions agnostic of the VC service (issuer, verifier..), it could make sense to export the VC service discovery from there.
